pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omniagent"
version = "0.1.0"
description = "OmniAgent production-ready multi-agent starter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<0.116",
  "uvicorn>=0.30,<0.31",
  "pydantic-settings>=2.2,<3",
  "httpx>=0.27,<0.28",
  "langgraph>=0.2,<0.3",
  "langsmith>=0.1,<0.2",
  "prometheus-client>=0.20,<0.21",
  "neo4j>=5.23,<6",
  "networkx>=3.3,<4",
  "rank-bm25>=0.2.2,<0.3",
  "numpy>=1.26,<3",
  "celery>=5.4,<6",
  "redis>=5.2,<6",
  "streamlit>=1.44,<2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
]

[tool.setuptools.packages.find]
include = ["backend*"]
exclude = ["frontend*", "tests*", "docs*", "scripts*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
